% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFactorLoadingModel.R
\name{applyFactorLoadingModel}
\alias{applyFactorLoadingModel}
\title{Applying factor loading model in parallel}
\usage{
applyFactorLoadingModel(mydata, solution, factorNames, faMethodScores)
}
\arguments{
\item{mydata}{A dataframe containing the all the initial analytical variables}

\item{solution}{A list containing the factor scores, previously calculated on the sample dataset}

\item{factorNames}{A vector containing the names of the factors}

\item{faMethodScores}{A string describing the method used for calculating the factors}
}
\description{
This function is used internally for applying the factor loading model inside the function; applyModels
}

