% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyModels.R
\name{applyModels}
\alias{applyModels}
\title{Applying all MachineLearneR models in parallel}
\usage{
applyModels(i, files, analyticalVariables, selectedMissingData,
  selectedNormalization, metaVariables, selectedTransformation,
  selectedStandardization, splitCol, classifierClass, removeCata, factorList,
  faMethodScores, selectedAverage, removeOutliers, controlVariable,
  controlValue, classModel, skew, kurto, selectedTrainingSize, createPlots,
  factors)
}
\arguments{
\item{i}{An integer, specifying nth parallel iteration}

\item{files}{A vector containing the filenames of the subsets}

\item{analyticalVariables}{A vector containing the names of the variables in the dataset that should be analysed}

\item{selectedMissingData}{A string specifying the method for imputing missing data}

\item{selectedNormalization}{A string specifying the method for normalizing the data}

\item{metaVariables}{A vector containing the names of the variables which should not be analysed, but should still remain in the dataset as describing variables}

\item{selectedTransformation}{A string specifying the method for transforming the data}

\item{selectedStandardization}{A string specifying the method for standardizing the data}

\item{splitCol}{A string specifying the name of the variable on which the dataset is splitted in subsets}

\item{classifierClass}{A string specifying the name of the variable on which the dataset should be classified}

\item{removeCata}{A boolean indicating if categorical variables should be removed from the dataset}

\item{factorList}{A list containing the factor scores, previously calculated on the sample dataset}

\item{faMethodScores}{A string describing the method used for calculating the factors}

\item{selectedAverage}{A string describing the basis for data normalization, can be mean, median or modus}

\item{removeOutliers}{A boolean indicating if outliers should be removed in data normalization}

\item{controlVariable}{A string specifying the variable on which the data should be normalized}

\item{controlValue}{A string specifying the value of the controlVariable on which the data should be normalized}

\item{classModel}{A randomforest fit object, to be applied to the data}

\item{skew}{A list containing the sample-calculated skewness per variable}

\item{kurto}{A list containing the sample-calculated kurtosis per variable}

\item{selectedTrainingSize}{An integer specifying the percentage of the dataset that should be used as trainingset for training the random forest model}

\item{createPlots}{A boolean indicating if plots should be generated}

\item{factors}{A vector containing the names of the sample-calculated factors, to be used as new analytical variables}
}
\description{
This function is used internally inside the \code{\link{ML}} function to apply all sample-generated models on the entire dataset
}

