% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlParameters.R
\name{ControlParameters}
\alias{ControlParameters}
\title{Checking parameters}
\usage{
ControlParameters(Parameters, bigdatanames)
}
\arguments{
\item{Parameters}{A vector containing the names of al the parameters used inside the ML function}

\item{bigdatanames}{A vector containing the names of al the variables in the input dataset}
}
\description{
This function is used inside the \code{\link{ML}} function to check if the given input parameters are in the correct format
}

