% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationBymissing.R
\name{correlationByMissing}
\alias{correlationByMissing}
\title{Function for finding NA values}
\usage{
correlationByMissing(x, listRm, mydata, analyticalVariables)
}
\arguments{
\item{x}{a string specifying the name of the current variable under investigation}

\item{mydata}{a dataframe containing the data to be analysed}

\item{analyticalVariables}{a vector containing the names of the variables to be analysed}

\item{listRM}{a list containing the variables to be removed from the dataset}
}
\description{
This function is used internally inside the function; eliminateHighestMissingVariableInCorrelationCutoff to find the NA values per variable
}

