% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClassModel.R
\name{createClassModel}
\alias{createClassModel}
\title{Using the random forests classifier}
\usage{
createClassModel(mydata, selectedTrainingSize, classifierClass,
  analyticalVariables, cores, createPlots, multiThreadFase, fit, parallelIter)
}
\arguments{
\item{mydata}{A dataframe}

\item{selectedTrainingSize}{A number between 1-100, describing the size of the training set as a percentage of the total dataset. Default is 80}

\item{classifierClass}{A string, containing the class to classify}

\item{analyticalVariables}{A character vector, containing the names of the variables in the dataset to analyze.}

\item{cores}{A number, describing the number of cores to use in parallel processing.}
}
\description{
Inside the \code{\link{ML}} function this function is used to perform a random forests classification on a given data frame. It creates a Test and a Training data set from the mydata input dataframe, and trains and applies a random forests classifier to predict the data.
}
\examples{
require("data.table")
data <- fread("data.csv")
data <- data.frame(data)
classModel(mydata=data, selectedTrainingSize = 90,classifierClass="ExampleClass", analyticalVariables = names(mydata), cores=6)
classModel(mydata=data, classifierClass="ExampleClass", analyticalVariables = names(mydata), cores=8)
}

