% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFLModel.R
\name{createFactorLoadingModel}
\alias{createFactorLoadingModel}
\title{Creating the factor loading model}
\usage{
createFactorLoadingModel(mydata, analyticalVariables,
  selectedNumberOfDimensions, selectedDataReductionMethod, removeCata, faRotate,
  faMethodScores, pcMethod, selectedCorrelationCutoff)
}
\arguments{
\item{mydata}{a dataframe containing the data to be analysed}

\item{analyticalVariables}{A vector containing the names of the variables in the dataset that should be analysed}

\item{selectedNumberOfDimensions}{the number of dimensions to be created, the number of produced factors. If 0, the number of dimensions is calculated by using the eigenvalues}

\item{removeCata}{A boolean indicating if categorical variables should be removed from the dataset}

\item{faRotate}{A string specifying the method to be used in rotation, default: oblimin}

\item{faMethodScores}{A string describing the method used for calculating the factors}

\item{pcMethod}{A string specifying the factoring method the be used}

\item{selectedCorrelationCutoff}{A number in range 0-1 specifying the maximum ammount of correlation between variables}
}
\description{
Inside the ML function this function creates a factor loading model, in order to reduce the ammount of dimensions from the original dataset. This function can be used seperately to generate a factor loading model for any given dataframe
}
\examples{
require("data.table")
data <- fread("data.csv")
data <- data.frame(data)
createFactorLoadingModel(mydata=data, analyticalVariables=c("var1", "var2"), removeCata=T, pcMethod='wls', selectedCorrelationCutoff=0.90)
}

