% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulationModel.R
\name{dMM}
\alias{dMM}
\title{Data manipulation}
\usage{
dMM(mydata, analyticalVariables, selectedNormalization, selectedAverage,
  selectedTransformation, selectedStandardization, removeOutliers, splitCol,
  controlVariable, controlValue, multiThreadFase, kurto, skew)
}
\arguments{
\item{mydata}{a dataframe containing the data to be analysed}

\item{analyticalVariables}{A vector containing the names of the variables in the dataset that should be analysed}

\item{selectedNormalization}{A string specifying the method for normalizing the data}

\item{selectedAverage}{A string describing the basis for data normalization, can be mean, median or modus}

\item{selectedTransformation}{A string specifying the method for transforming the data}

\item{selectedStandardization}{A string specifying the method for standardizing the data}

\item{removeOutliers}{A boolean indicating if outliers should be removed in data normalization}

\item{splitCol}{A string specifying the name of the variable on which the dataset is splitted in subsets}

\item{controlVariable}{A string specifying the variable on which the data should be normalized}

\item{controlValue}{A number specifying the value of the controlVariable on which the data should be normalized}

\item{multiThreadFase}{A boolean indicating if this function is excecuted in parallel processing}

\item{kurto}{A list containing the sample-calculated kurtosis per variable, only required if multiThreadFase=TRUE}

\item{skew}{A list containing the sample-calculated skewness per variable, only required if multiThreadFase=TRUE}
}
\description{
Inside the ML function this function is responsible for manipulating the data in the dataset. This function can be used seperately to manipulate data for any given dataframe.
}
\examples{
require("data.table")
data <- fread("data.csv")
data <- data.frame(data)
dMM(mydata=data, analyticalVariables=c("var1", "var2"), selectedNormalization=normalize.onClass, controlVariable='var2', controlValue=22, splitCol='var1')
}

