% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminateHighestMissingVariableInCorrelationCutoff.R
\name{eliminateHighestMissingVariableInCorrelationCutoff}
\alias{eliminateHighestMissingVariableInCorrelationCutoff}
\title{Removing the variables containing the most NA values}
\usage{
eliminateHighestMissingVariableInCorrelationCutoff(mydata, analyticalVariables,
  selectedCorrelationCutoff, removeCata, corData = NULL)
}
\arguments{
\item{mydata}{a dataframe containing the data to be analysed}

\item{analyticalVariables}{A vector containing the names of the variables in the dataset that should be analysed}

\item{selectedCorrelationCutoff}{A number in range 0-1 specifying the maximum ammount of correlation between variables}

\item{removeCata}{A boolean indicating if categorical variables should be removed from the dataset}

\item{corData}{A matrix containing information about the ammount of correlation between variables, if NULL, the function calculates this by itsself.}
}
\description{
This function is used internally inside the function; \code{\link{ExcludedVariablesModel}} to remove the variables from the dataset containing to many NA values.
}

