% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factor.analysis.R
\name{factorAnalysis}
\alias{factorAnalysis}
\title{Applying factor analysis}
\usage{
factorAnalysis(mydata, selectedNumberOfDimensions, covList, faRotate,
  faMethodScores, pcMethod)
}
\arguments{
\item{mydata}{a dataframe containing the data to be analysed}

\item{selectedNumberOfDimensions}{the number of dimensions to be created, the number of produced factors. If 0, the number of dimensions is calculated by using the eigenvalues}

\item{covList}{A list with the components cov, center, and n.obs.}

\item{faRotate}{A string specifying the method to be used in rotation, default: oblimin}

\item{faMethodScores}{A string describing the method used for calculating the factors}

\item{pcMethod}{A string specifying the factoring method the be used.}
}
\description{
Inside the \code{\link{createFactorLoadingModel}} function this function is used to perform factor analysis on a given data frame.
}

