% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filesplit.R
\name{Filesplit}
\alias{Filesplit}
\title{Function writing a data frame to an .Rdata File}
\usage{
Filesplit(xSet, splitCol)
}
\arguments{
\item{xSet}{A dataframe}

\item{splitCol}{A String, specifying a column of the xSet dataframe, which is used to generate the filename.}
}
\value{
fileName A String, specifying the name of the file created by the function.
}
\description{
\code{Filesplit} generates a number, based on the splitCol parameter, which it used to create an .Rdata file of the input data frame, xSet, which it saves.
}
\examples{
FileSplit(mydata, splitCol="Col1")
}

