% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInput.R
\name{GetInput}
\alias{GetInput}
\title{Function for loading input for the ML package}
\usage{
GetInput(mydata, dir, subsetCutoff, splitCol, usedCores, sampleSize,
  selectedGarbageCutoff, selectedCorrelationCutoff, removeCata, metaVariables,
  controlVariable, classifierClass)
}
\arguments{
\item{mydata}{A dataframe}

\item{dir}{A location specifying the working directory. This is where the .Rdata files of the subsets will be saved.}

\item{subsetCutoff}{A number, specifying the minimal number of records that must be in a subset.}

\item{splitCol}{A string, specifying the column on which the data set should be splitted.}

\item{usedCores}{A number, specifing the number of cores to be used in parallel processing, this cannot be more than the maximum number of cores availible on the system.}

\item{sampleSize}{A number, in range 0-1, specifying the fraction of the mydata dataframe to be used as sample data set. Default is 0.05 (5\%)}

\item{selectedGarbageCutoff}{A number, in range 0-1, specifying the fraction of records in a variable(column) that is accepted to be NA. Default is 0.10(10\%).}

\item{selectedCorrelationCutoff}{A number, in range 0-1, specifying the fraction of the maximum accepted ammount of correlation between two variables.}

\item{removeCata}{A boolean, enabling/disabling the removal of catagorical variables(columns) from the mydata data frame. Default is TRUE, catagorical varibiables are removed.}

\item{metaVariables}{A character vector, containing the names of the variables(columns) that should NOT be analyzed, but should remain in de mydata data frame as describing variables.}
}
\value{
\describe{
A list containing:
  \item{samp}{The sample data frame}
  \item{files}{the vector containing the names of the splitted .Rdata files}
  \item{splitCol}{The column on which the files are splitted}
  \item{analyticalVariables}{The vector containing the names of the columns to be analyzed}
  \item{cores}{A number, specifing the number of cores that was used during parallel processing}
  \item{garbagedf}{A data frame, containing all the variables that were removed by the \code{\link{ExcludedVariablesModel}} function.}
  \item{metaVariables}{A vector, containing the updated version of the metaVariables input parameter}
}
}
\description{
\code{GetInput} checks if the mydata input is a data frame or a file, if it is a file, the function uses fread from the package data.table to read the data into a data frame. After defining the mydata data frame, this function sets the number of cores to be used during parallel processing, based on the input parameters. Based on the sampleSize parameter, the function generates a stratified sample of the mydata data frame. The function \code{\link{ExcludedVariablesModel}} is called on the sample, which reduces the number of variables(columns). The \code{\link{ExcludedVariablesModel}} function returns a Vector with columnnames that should be analyzed, analyticalVariables, which this function uses to reduce the number of variables in de mydata data frame. In parallel processing, this function then splits the (big) mydata data frame, to smaller subsets in .Rdata files by calling the \code{\link{Filesplit}} function in combination with the \code{\link[plyr]{dlply}} function.
}
\examples{
GetInput(mydata=bigdata,dir="dir"~/Home",subsetCutoff=10000, usedCores=8,selectedGarbageCutoff=0.2, selectedCorrelationCutoff=0.80,removeCata=FALSE, metaVariables=c("Col1",Col2","Col3"))
}

