% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultipleImpMod.R
\name{impute.median}
\alias{impute.median}
\title{Imputation by median}
\usage{
impute.median(x)
}
\arguments{
\item{x}{A vector containing NA values}
}
\description{
Inside the \code{\link{cMIM}} function this function is responsible for imputing the missing values in the dataset by median. This function can be used separately to impute NA values for any given vector.
}

