% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultipleImpMod.R
\name{impute.modus}
\alias{impute.modus}
\title{Imputation by mode}
\usage{
impute.modus(x)
}
\arguments{
\item{x}{A vector containing NA values}
}
\description{
Inside the \code{\link{cMIM}} function this function is responsible for imputing the missing values in the dataset by mode. This function can be used separately to impute NA values for any given vector.
}

