% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.R
\name{ML}
\alias{ML}
\title{Main function for running the ML package}
\usage{
ML(bigdata, dir = default.dir, subsetCutoff = default.subsetCutoff,
  splitCol = default.splitCol, usedCores = default.usedCores,
  sampleSize = default.sampleSize, createPlots = default.createPlots,
  selectedGarbageCutoff = default.selectedGarbageCutoff,
  selectedCorrelationCutoff = default.selectedCorrelationCutoff,
  selectedMissingData = default.selectedMissingData,
  classifierClass = default.classifierClass,
  removeCata = default.removeCata, metaVariables = default.metaVariables,
  selectedNormalization = default.selectedNormalization,
  selectedTransformation = default.selectedTransformation,
  selectedAverage = default.selectedAverage,
  selectedStandardization = default.selectedStandardization,
  controlVariable = default.controlVariable,
  controlValue = default.controlValue,
  removeOutliers = default.removeOutliers,
  selectedDataReductionMethod = default.selectedDataReductionMethod,
  selectedNumberOfDimensions = default.selectedNumberOfDimensions,
  faRotate = default.faRotate, faMethodScores = default.faMethodScores,
  pcMethod = default.pcMethod,
  selectedTrainingSize = default.selectedTrainingSize)
}
\arguments{
\item{bigdata}{A dataframe or flatfile containing a dataset.}

\item{dir}{A location specifying the working directory. Default is \code{\link{getwd}}}

\item{subsetCutoff}{A number, specifying the minimal number of records that must be in a subset. Default is 10000}

\item{splitCol}{A string, specifying the column on which the data set should be splitted.}

\item{usedCores}{A number, specifing the number of cores to be used in parallel processing.}

\item{sampleSize}{A number, in range 0-1, specifying the fraction of the mydata dataframe to be used as sample data set. Default is 0.05 (5\%).}

\item{selectedGarbageCutoff}{A number, in range 0-1, specifying the fraction of records in a variable(column) that is accepted to be NA. Default is 0.10(10\%).}

\item{selectedCorrelationCutoff}{A number, in range 0-1, specifying the fraction of the maximum accepted ammount of correlation between two variables.}

\item{selectedMissingData}{A String, specifying the method to be used when calling the imputation model function; cMIM which should be one of the following methods:
\itemize{
   \item "CWD"
   \item "impute.modus"
   \item "impute.median"
   \item "impute.mean"
   \item "onclass.modus"
   \item "onclass.median"
   \item "onclass.mean"
   \item "regression"
 }}

\item{removeCata}{A boolean, enabling/disabling the removal of catagorical variables(columns) from the mydata data frame. Default is TRUE, catagorical variables are removed.}

\item{rm}{nnet, rpart, survival}
}
\description{
\code{ML}
}

