% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulationModel.R
\name{normalize}
\alias{normalize}
\title{Normalizing data the data}
\usage{
normalize(currentVariable, currentVariableValues, averageData)
}
\arguments{
\item{currentVariable}{temp}

\item{currentVariableValues}{temp}

\item{averageData}{temp}
}
\description{
Inside the \code{\link{dMM}} function this function is responsible for normalizing the dataset.
}

