% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultipleImpMod.R
\name{onclass}
\alias{onclass}
\title{Imputation by class}
\usage{
onclass(df, impClass, func)
}
\arguments{
\item{df}{A data frame containing NA values}

\item{impClass}{A variable on which the records should be aggregated}

\item{func}{A function name, specifying the method to be used to impute the values}
}
\description{
Inside the \code{\link{cMIM}} function this function is responsible for imputing the missing values in the dataset by class. This function can be used separately to impute NA values for any given data frame.
}

