% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Principalcomponentanalysis.R
\name{Principalcomponentanalysis}
\alias{Principalcomponentanalysis}
\title{Perfoming PCA}
\usage{
Principalcomponentanalysis(mydata, selectedNumberOfDimensions, covList,
  faMethodScores)
}
\description{
temp
}

