% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcludedVariablesModel.R
\name{removeCategorical}
\alias{removeCategorical}
\title{Removing Catagorical}
\usage{
removeCategorical(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Inside the \code{\link{ExcludedVariablesModel}} function this function is responsible for
}

