% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcludedVariablesModel.R
\name{removeMissing}
\alias{removeMissing}
\title{Removing missing}
\usage{
removeMissing(x, selectedGarbageCutoff)
}
\arguments{
\item{x}{A vector}
}
\description{
Inside the \code{\link{ExcludedVariablesModel}} function this function is responsible for
}

