% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcludedVariablesModel.R
\name{removeNonNumeric}
\alias{removeNonNumeric}
\title{Removing Non-numeric}
\usage{
removeNonNumeric(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Inside the \code{\link{ExcludedVariablesModel}} function this function is responsible for
}

