% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulationModel.R
\name{robustZscore}
\alias{robustZscore}
\title{RobustZscore standardisation.}
\usage{
robustZscore(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Inside the \code{\link{dMM}} function this function is responsible for standardizing the variables based on their robust Z score.
}

