% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulationModel.R
\name{transformData}
\alias{transformData}
\title{Transforming data the data}
\usage{
transformData(x, skew)
}
\arguments{
\item{x}{A vector}

\item{skew}{A vector}
}
\description{
Inside the \code{\link{dMM}} function this function is responsible for log transforming the variables based on their skewness.
}

