% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClassModel.R
\name{variableImportancePlot}
\alias{variableImportancePlot}
\title{Plotting Variable importance}
\usage{
variableImportancePlot(fit, classifierClass, analyticalVariables, parallelIter)
}
\description{
Inside the \code{\link{createClassModel}} function this function is responsible for
}

