% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcludedVariablesModel.R
\name{ExcludedVariablesModel}
\alias{ExcludedVariablesModel}
\title{Function for reducing the number of variables(columns) in a data frame}
\usage{
ExcludedVariablesModel(mydata, analyticalVariables, selectedGarbageCutoff,
  selectedCorrelationCutoff, removeCata)
}
\arguments{
\item{mydata}{a dataframe containing the data to be analysed}

\item{analyticalVariables}{A vector containing the names of the variables in the dataset that should be analysed}
}
\description{
blabalba
}

