% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInput.R
\name{Filesplit}
\alias{Filesplit}
\title{Writing data frame subsets to an .Rdata File}
\usage{
Filesplit(xSet, splitCol)
}
\arguments{
\item{xSet}{A dataframe}

\item{splitCol}{A String, specifying a column of the xSet dataframe, which is used to generate the filename.}
}
\description{
This function is used inside the \code{\link{GetInput}} function to split a data frame into subsets by using \code{\link[plyr]{dlply}}.\code{Filesplit} generates a number, based on the splitCol parameter, which it uses to create an .Rdata file of the input data frame, xSet, which it saves.
}
\examples{
FileSplit(mydata, splitCol="Col1")
}

